Added test scene - InstanceTestScene - demonstrating power of instancing
this test-scene shows that the implementation of instancing seems to
work. the basic idea is still to use one instance of an intersectable
and then just pass a reference of this intersectable to an instance.
modifications of similar intersectables are then described by
homogeneous tranformations instead manipulating the geometries, i.e.
intersectables, directly.